feature/offset_type_optimize

See merge request kinectpro/whattowear!71

+++ app/src/main/java/com/kinectpro/whattowear/data/model/response/WeatherData.kt
@@ -11,7 +11,7 @@ data class DarkSkyWeather(
     val latitude: Double,
     val longitude: Double,
     val daily: Daily,
-    val offset: Long
+    val offset: Float
 )
 
 data class Daily(
